requirements.txt
Django==3.2.8
psycopg2==2.9.1
djangorestframework==3.12.4
requests>=2.26.0
python-dotenv==0.19.1
pytest-django
pytest